third_party/blink/web_tests/platform/win11-arm64/http/tests/images/image-with-dpr-natural-dimensions-expected.txt

This is a testharness.js-based test.
FAIL Test srcset and sizes with a meta viewport assert_equals: expected 25 but got 0
Harness: the test ran to completion.
